fix: 개인정보 저장 및 불러오기 진행 시 암호화, 복호화 진행하도록 수정

+++ src/main/java/com/purchase/hanghae99/user/UserServiceImpl.java
@@ -22,6 +22,7 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.core.Authentication;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -30,6 +31,7 @@ import java.time.Duration;
 import java.util.HashMap;
 import java.util.Map;
 
+import static com.purchase.hanghae99.common.AesUtils.*;
 import static com.purchase.hanghae99.common.CustomCookieManager.*;
 import static com.purchase.hanghae99.common.exception.ExceptionCode.*;
 
@@ -43,12 +45,11 @@ public class UserServiceImpl implements UserService{
     private final EmailService emailService;
     private final JwtProvider jwtProvider;
     private final CustomCookieManager cookieManager;
-
     private final RedisService redisService;
 
     @Override
     @Transactional
-    public ResUserCreateDto createUser(ReqUserCreateDto reqDto) {
+    public ResUserCreateDto createUser(ReqUserCreateDto reqDto) throws Exception {
         checkDuplicateEmail(reqDto.getEmail());
         checkDuplicatePhoneNumber(reqDto.getPhoneNumber());
 
@@ -63,16 +64,22 @@ public class UserServiceImpl implements UserService{
     public ResUserInfoDto readUser(Long userId) {
         return userRepository.findById(userId)
                 .filter(x -> x.getRole().equals(UserRole.CERTIFIED_USER))
-                .map(ResUserInfoDto::fromEntity)
+                .map(user -> {
+                    try {
+                        return ResUserInfoDto.fromEntity(user);
+                    } catch (Exception e) {
+                        throw new BusinessException(DECODING_ERROR);
+                    }
+                })
                 .orElseThrow(() -> new BusinessException(NOT_FOUND_USER));
     }
 
     @Override
     @Transactional
-    public ResUserUpdateDto updateUserInfo(Long userId, ReqUserInfoUpdateDto reqDto) {
+    public ResUserUpdateDto updateUserInfo(Long userId, ReqUserInfoUpdateDto reqDto) throws Exception {
         User savedUser = userRepository.findById(userId)
                 .orElseThrow(() -> new BusinessException(NOT_FOUND_USER));
-        savedUser.updateUserInfo(reqDto.getPhoneNumber(), reqDto.getAddress());
+        savedUser.updateUserInfo(aesCBCEncode(reqDto.getPhoneNumber()), aesCBCEncode(reqDto.getAddress()));
         return ResUserUpdateDto.fromEntity(userRepository.save(savedUser));
     }
 
@@ -91,11 +98,11 @@ public class UserServiceImpl implements UserService{
 
     @Override
     @Transactional
-    public ResEmailDto updateEmailVerification(Long userId, String userStr) {
+    public ResEmailDto updateEmailVerification(Long userId, String userStr) throws Exception {
         User savedUser = userRepository.findById(userId)
                 .orElseThrow(() -> new BusinessException(NOT_FOUND_USER));
 
-        if (!emailService.checkVerificationStr(savedUser.getEmail(), userStr)) {
+        if (!emailService.checkVerificationStr(aesCBCDecode(savedUser.getEmail()), userStr)) {
             return EmailDtoFactory.fail();
         }
 
@@ -106,10 +113,13 @@ public class UserServiceImpl implements UserService{
 
     @Override
     @Transactional
-    public void deleteUser(Long userId, ReqUserDeleteDto reqDto) {
+    public void deleteUser(Authentication authentication, Long userId, ReqUserDeleteDto reqDto) throws Exception {
         User savedUser = userRepository.findById(userId)
                 .orElseThrow(() -> new BusinessException(NOT_FOUND_USER));
 
+        String email = authentication.getName();
+
+        verifyAccessedUser(email, savedUser.getEmail());
         verifyPassword(reqDto.getPassword(), reqDto.getPassword2());
 
         userRepository.delete(savedUser);
@@ -117,8 +127,8 @@ public class UserServiceImpl implements UserService{
 
     @Override
     @Transactional
-    public ResLoginDto login(HttpServletResponse response, ReqLoginDto reqDto) {
-        User user = userRepository.findByEmail(reqDto.getEmail())
+    public ResLoginDto login(HttpServletResponse response, ReqLoginDto reqDto) throws Exception {
+        User user = userRepository.findByEmail(aesCBCEncode(reqDto.getEmail()))
                 .orElseThrow(() -> new BusinessException(NOT_FOUND_USER));
 
         if (!passwordEncoder.matches(reqDto.getPassword(), user.getPassword())) {
@@ -129,8 +139,10 @@ public class UserServiceImpl implements UserService{
             throw new BusinessException(UNCERTIFIED_EMAIL);
         }
 
+        String decodedEmail = aesCBCDecode(user.getEmail());
+
         Map<String, Object> claims = new HashMap<>();
-        claims.put(JwtProvider.EMAIL_CLAIM, user.getEmail());
+        claims.put(JwtProvider.EMAIL_CLAIM, decodedEmail);
 
         String accessToken = jwtProvider.createAccessToken(claims);
         String refreshToken = jwtProvider.createRefreshToken();
@@ -138,7 +150,7 @@ public class UserServiceImpl implements UserService{
         cookieManager.setCookie(response, accessToken, ACCESS_TOKEN, jwtProvider.getAccessTokenExpirationPeriod());
         cookieManager.setCookie(response, refreshToken, REFRESH_TOKEN, jwtProvider.getRefreshTokenExpirationPeriod());
 
-        redisService.setValues(user.getEmail(), refreshToken, Duration.ofMillis(jwtProvider.getRefreshTokenExpirationPeriod()));
+        redisService.setValues(decodedEmail, refreshToken, Duration.ofMillis(jwtProvider.getRefreshTokenExpirationPeriod()));
 
         return ResLoginDto.fromEntity(user);
     }
@@ -167,14 +179,14 @@ public class UserServiceImpl implements UserService{
         }
     }
 
-    private void checkDuplicateEmail(String email) {
-        if (userRepository.existsByEmail(email)) {
+    private void checkDuplicateEmail(String email) throws Exception {
+        if (userRepository.existsByEmail(aesCBCEncode(email))) {
             throw new BusinessException(ALREADY_REGISTERED_EMAIL);
         }
     }
 
-    private void checkDuplicatePhoneNumber(String phoneNumber) {
-        if (userRepository.existsByPhoneNumber(phoneNumber)) {
+    private void checkDuplicatePhoneNumber(String phoneNumber) throws Exception {
+        if (userRepository.existsByPhoneNumber(aesCBCEncode(phoneNumber))) {
             throw new BusinessException(ALREADY_REGISTERED_PHONE_NUMBER);
         }
     }
@@ -188,6 +200,12 @@ public class UserServiceImpl implements UserService{
     public void verifyPassword(String password, String password2) {
         if (!password.equals(password2)) {
             throw new BusinessException(INVALID_PASSWORD);
+        }
+    }
+
+    public void verifyAccessedUser(String authenticatedEmail, String emailOfUser) throws Exception {
+        if (!authenticatedEmail.equals(aesCBCDecode(emailOfUser))) {
+            throw new BusinessException(UNAUTHORIZED_ACCESS);
         }
     }
 }
